+++ src/main/gen/com/github/nthykier/debpkg/deb822/psi/Deb822FieldValuePair.java
@@ -4,13 +4,17 @@ package com.github.nthykier.debpkg.deb822.psi;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.ContributedReferenceHost;
+import com.intellij.psi.PsiReference;
 
-public interface Deb822FieldValuePair extends PsiElement {
+public interface Deb822FieldValuePair extends ContributedReferenceHost {
 
   @NotNull
   Deb822Field getField();
 
   @Nullable
   Deb822ValueParts getValueParts();
+
+  @Nullable PsiReference @NotNull [] getReferences();
 
 }