+++ src/main/gen/com/github/nthykier/debpkg/deb822/psi/Deb822Value.java
@@ -8,7 +8,6 @@ import com.intellij.psi.PsiReference;
 
 public interface Deb822Value extends PsiElement {
 
-  @Nullable
-  PsiReference getReference();
+  @Nullable PsiReference getReference();
 
 }