+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - A quick fix to the Deb822 copyright format that converts directories into
   a directory + wildcard as directories themselves match nothing.
 - Support for IDEA 2020.3.
+- Valid paths without wildcards in the `Files` field in `debian/copyright`
+  are now turned into references (enabling "jump to" and refactoring).
 
 ### Changed
 - Parse `[ Name ]:` as an alternative to `[ Name ]` token in `debian/changelog`.